Adjust the year

adjust_year() adjusts the year of x to value.

adjust_year(x, value, ...)

# S3 method for Date
adjust_year(x, value, ..., day_nonexistent = "last-time")

# S3 method for POSIXt
adjust_year(
  x,
  value,
  ...,
  day_nonexistent = "last-time",
  dst_nonexistent = "roll-forward",
  dst_ambiguous = "earliest"
)

# S3 method for clock_zoned_time_point
adjust_year(
  x,
  value,
  ...,
  day_nonexistent = "last-time",
  dst_nonexistent = "roll-forward",
  dst_ambiguous = "earliest"
)

# S3 method for clock_naive_time_point
adjust_year(x, value, ..., day_nonexistent = "last-time")

# S3 method for clock_calendar
adjust_year(x, value, ..., day_nonexistent = "last-time")

Arguments

x
[Date / POSIXct / POSIXlt]

A date-time vector.

value
[integer]

An integer vector containing the value to adjust to.

day_nonexistent
[character(1)]

Control the behavior when a nonexistent day is generated. This only happens when adjusting years, months, or days.

"last-time": Adjust to the last possible time of the current month.

"first-time": Adjust to the first possible time of the following month.

"last-day": Adjust to the last day of the current month. For date-times, the sub-daily components are kept.

"first-day": Adjust to the first day of the following month. For date-times, the sub-daily components are kept.

"NA": Replace the nonexistent date with NA.

"error": Error on nonexistent dates.

Examples

x <- as.Date(c("1970-01-01", "1971-01-01"))
adjust_year(x, 1972)

#> [1] "1972-01-01" "1972-01-01"

# Leap day
x <- as.Date("1972-02-29")

# "1973-02-29" doesn't exist. By default this rolls back to the
# the last day in that month.
adjust_year(x, 1973)

#> [1] "1973-02-28"

# But you can adjust that behavior
adjust_year(x, 1973, day_nonexistent = "first-day")

#> [1] "1973-03-01"

adjust_year(x, 1973, day_nonexistent = "NA")

#> [1] NA

# It is possible to adjust into a daylight savings time gap.
# Due to a change in how daylight savings was handled, the Pacific
# island of Samoa skipped all of "2011-12-30" and went straight to the
# 31st. This is considered a "nonexistent" time due to DST.
x <- as.POSIXct("2010-12-30 02:00:00", "Pacific/Apia")

# So "2011-12-30 02:00:00" doesn't exist, and by default we "roll forward"
# to the next possible time, which is "2011-12-31 00:00:00".
adjust_year(x, 2011)

#> [1] "2011-12-31 +14"

# But you can adjust that too
adjust_year(x, 2011, dst_nonexistent = "roll-backward")

#> [1] "2011-12-29 23:59:59 -10"

adjust_year(x, 2011, dst_nonexistent = "shift-forward")

#> [1] "2011-12-31 02:00:00 +14"