Common date-time formats

These functions are pre-generated formats created from combining multiple tokens with tkns(). You can use them directly as formats in format() for civil classes or Date and POSIXct objects.

fmt_ym(..., year = tkn_year(), month = tkn_month(), sep = "-")

fmt_ymd(
  ...,
  year = tkn_year(),
  month = tkn_month(),
  day = tkn_day(),
  sep = "-"
)

fmt_hms(
  ...,
  hour = tkn_hour(),
  minute = tkn_minute(),
  second = tkn_second(),
  sep = ":"
)

fmt_ymd_hms(..., ymd = fmt_ymd(), hms = fmt_hms(), sep = " ")

fmt_ymd_hms_zoned(
  ...,
  ymd = fmt_ymd(),
  hms = fmt_hms(),
  sep = " ",
  colon = TRUE,
  zone_name = TRUE
)

fmt_week_date()

fmt_ordinal_date()

fmt_iso_week_date()

fmt_iso_date()

fmt_iso_datetime()

fmt_local_year_month()

fmt_local_date()

fmt_local_datetime()

fmt_local_nano_datetime()

fmt_zoned_datetime(..., zone_name = TRUE)

fmt_zoned_nano_datetime(..., zone_name = TRUE)

Arguments

...
These dots are for future extensions and must be empty.

year
[character(1)]

The token to use to represent the year.

month
[character(1)]

The token to use to represent the month.

sep
[character(1)]

The separator to use between tokens.

day
[character(1)]

The token to use to represent the day of the month.

hour
[character(1)]

The token to use to represent the hour.

minute
[character(1)]

The token to use to represent the minute.

second
[character(1)]

The token to use to represent the second.

ymd
[character(1)]

The format to use to represent the year, month, and day.

hms
[character(1)]

The format to use to represent the hour, minute, and second.

colon
[logical(1)]

If TRUE, a ":" will be added between the hours and minutes of the zone offset.

zone_name
[logical(1)]

If TRUE, an extension to the ISO-8601 standard datetime format is used. This extension adds the time zone name or abbreviation to the end of the standard format as [%Z].

This is most useful with the format() method for civil's zoned datetimes, as it can generate a textual representation of a zoned datetime that can be unambiguously parsed back into the original object.

Examples

fmt_ymd_hms()

#> [1] "%Y-%m-%d %H:%M:%S"

x <- as.Date("2018-12-31")

# ISO weeks are counted differently from normal weeks!
format(x, fmt_week_date())

#> [1] "2018-W52-1"

format(x, fmt_iso_week_date())

#> [1] "2019-W01-1"

# Date that is "ambiguous" without more information about the DST offset
x <- zoned_datetime(
  1970, 10, 25, 01, 30,
  zone = "America/New_York",
  dst_ambiguous = c("earliest", "latest")
)
x

#> <civil_datetime<America/New_York>[2]>
#> [1] "1970-10-25 01:30:00-04:00" "1970-10-25 01:30:00-05:00"

# The ISO datetime format is incomplete because we don't know which
# time zone we came from
format(x, format = fmt_iso_datetime())

#> [1] "1970-10-25T01:30:00-0400" "1970-10-25T01:30:00-0500"

# The default print method for POSIXct isn't much better, because the
# time zone abbreviation of "EST" is used by multiple different time zones,
# not just "America/New_York".
format(as.POSIXct(x), usetz = TRUE)

#> [1] "1970-10-25 01:30:00 EDT" "1970-10-25 01:30:00 EST"

# The only way to fully capture the date-time information is to use the
# zoned date-time format that prints both the offset and the zone name.
format(x, format = fmt_ymd_hms_zoned())

#> [1] "1970-10-25 01:30:00-04:00[America/New_York]"
#> [2] "1970-10-25 01:30:00-05:00[America/New_York]"

# The default format method for zoned datetimes uses this extended format
format(x)

#> [1] "1970-10-25 01:30:00-04:00[America/New_York]"
#> [2] "1970-10-25 01:30:00-05:00[America/New_York]"