Tokens for date-time formatting

This page contains a number of tokens that can be used for date-time formatting through format(format = ). All tokens are supported for formatting a civil class (such as year_month_day() or zoned_datetime()). Most tokens are supported for formatting a POSIXct or Date, but some may be specific to the formatting implementation used by civil.

Calling any of these functions simply returns the corresponding string for that token.

There are also pre-generated formats for the most common cases, like fmt_ymd_hms(). If you need to generate a custom format, multiple tokens can be combined together using tkns(). You can also use glue::glue(), which can often be more readable.

tkn_weekday_name(..., abbreviate = FALSE)

tkn_month_name(..., abbreviate = FALSE)

tkn_date_and_time(..., locale_era = FALSE)

tkn_century(..., locale_era = FALSE)

tkn_day(..., pad = "zero", locale_num = FALSE)

tkn_iso_year(..., short = FALSE)

tkn_hour(..., twelve = FALSE, locale_num = FALSE)

tkn_day_of_year()

tkn_month(..., locale_num = FALSE)

tkn_minute()

tkn_newline()

tkn_am_pm()

tkn_time(..., twelve = FALSE)

tkn_second(..., locale_num = FALSE)

tkn_tab()

tkn_iso_weekday(..., locale_num = FALSE)

tkn_weeknum(..., locale_num = FALSE, week_start_sunday = TRUE)

tkn_iso_weeknum(..., locale_num = FALSE)

tkn_weekday(..., locale_num = FALSE)

tkn_year(..., short = FALSE)

tkn_offset(..., colon = TRUE)

tkn_zone()

tkn_percent()

Arguments

...
These dots are for future extensions and must be empty.

abbreviate
[logical(1)]

Should the abbreviated version of this token be used?

locale_era
[logical(1)]

Should the alternative locale-dependent era format be used? This prepends an E before the standard token (i.e. %EY rather than %Y).

pad
[character(1)]

One of "zero" or "space", specifying how to pad days of the month that have a single digit. For example, "1" would be padded to "01" or " 1".

locale_num
[logical(1)]

Should the alternative locale-dependent number format be used? This prepends an O before the standard token (i.e. %OH rather than %H).

short
[logical(1)]

If TRUE, a two digit year will be printed.

If FALSE, a year with at least 4 digits will be printed.

twelve
[logical(1)]

If TRUE, a 12-hour clock will be used.

If FALSE, a 24-hour clock will be used.

week_start_sunday
[logical(1)]

If TRUE, the week starts on Sunday, which is the US convention.

If FALSE, the week starts on Monday, which is the UK convention.

colon
[logical(1)]

If TRUE, a ":" will be added between the hours and minutes of the zone offset.

Examples

# Tokens are just strings that start with a `%`.
tkn_day()

#> [1] "%d"

tkn_weekday()

#> [1] "%w"

# Multiple tokens can be combined with `tkns()` to create custom formats
hms <- tkns(tkn_hour(), tkn_minute(), tkn_second(), sep = ":")

# Formats created by tokens can be used in the format method of civil
# classes and base R date-time classes
x <- naive_datetime(2019, 01, 01, 14, 30, 59)
x

#> <naive_datetime[1]>
#> [1] "2019-01-01 14:30:59"

format(x, format = hms)

#> [1] "14:30:59"

format(x, format = tkn_time(twelve = TRUE))

#> [1] "02:30:59 PM"

x_ct <- as.POSIXct(x, tz = "America/New_York")

format(x_ct, format = hms)

#> [1] "14:30:59"