Zoned date-time arithmetic

Zoned arithmetic involves adding or subtracting units of time from a datetime that has a time zone attached. This means that all of the complexities of daylight savings time and nonexistent days have to be handled after each individual arithmetic operation. The alternative is naive arithmetic, which only deals with the complexities of time zones once after all arithmetic has been performed.

Zoned arithmetic is usually fine for additions of singular units of time. If you want to add multiple periods, consider switching to naive arithmetic.

The following add periods of time. Periods are units of time that do not have a fixed constant duration (i.e. a "month" may be 30 or 31 days).

add_years()

add_months()

add_days()

The following add durations of time. Durations are fixed units of time based on a set number of seconds.

add_hours()

add_minutes()

add_seconds()

When nonexistent or ambiguous date-times are landed on, the day_nonexistent, dst_nonexistent, and dst_ambiguous options are consulted to resolve any issues.

# S3 method for Date
add_years(x, n, ..., day_nonexistent = "last-time")

# S3 method for POSIXt
add_years(
  x,
  n,
  ...,
  day_nonexistent = "last-time",
  dst_nonexistent = NULL,
  dst_ambiguous = NULL
)

# S3 method for civil_zoned
add_years(
  x,
  n,
  ...,
  day_nonexistent = "last-time",
  dst_nonexistent = NULL,
  dst_ambiguous = NULL
)

# S3 method for Date
subtract_years(x, n, ..., day_nonexistent = "last-time")

# S3 method for POSIXt
subtract_years(
  x,
  n,
  ...,
  day_nonexistent = "last-time",
  dst_nonexistent = NULL,
  dst_ambiguous = NULL
)

# S3 method for civil_zoned
subtract_years(
  x,
  n,
  ...,
  day_nonexistent = "last-time",
  dst_nonexistent = NULL,
  dst_ambiguous = NULL
)

# S3 method for Date
add_months(x, n, ..., day_nonexistent = "last-time")

# S3 method for POSIXt
add_months(
  x,
  n,
  ...,
  day_nonexistent = "last-time",
  dst_nonexistent = NULL,
  dst_ambiguous = NULL
)

# S3 method for civil_zoned
add_months(
  x,
  n,
  ...,
  day_nonexistent = "last-time",
  dst_nonexistent = NULL,
  dst_ambiguous = NULL
)

# S3 method for Date
subtract_months(x, n, ..., day_nonexistent = "last-time")

# S3 method for POSIXt
subtract_months(
  x,
  n,
  ...,
  day_nonexistent = "last-time",
  dst_nonexistent = NULL,
  dst_ambiguous = NULL
)

# S3 method for civil_zoned
subtract_months(
  x,
  n,
  ...,
  day_nonexistent = "last-time",
  dst_nonexistent = NULL,
  dst_ambiguous = NULL
)

# S3 method for Date
add_years_and_months(x, years, months, ..., day_nonexistent = "last-time")

# S3 method for POSIXt
add_years_and_months(
  x,
  years,
  months,
  ...,
  day_nonexistent = "last-time",
  dst_nonexistent = NULL,
  dst_ambiguous = NULL
)

# S3 method for civil_zoned
add_years_and_months(
  x,
  years,
  months,
  ...,
  day_nonexistent = "last-time",
  dst_nonexistent = NULL,
  dst_ambiguous = NULL
)

# S3 method for Date
subtract_years_and_months(x, years, months, ..., day_nonexistent = "last-time")

# S3 method for POSIXt
subtract_years_and_months(
  x,
  years,
  months,
  ...,
  day_nonexistent = "last-time",
  dst_nonexistent = NULL,
  dst_ambiguous = NULL
)

# S3 method for civil_zoned
subtract_years_and_months(
  x,
  years,
  months,
  ...,
  day_nonexistent = "last-time",
  dst_nonexistent = NULL,
  dst_ambiguous = NULL
)

# S3 method for Date
add_weeks(x, n, ...)

# S3 method for POSIXt
add_weeks(x, n, ..., dst_nonexistent = NULL, dst_ambiguous = NULL)

# S3 method for civil_zoned
add_weeks(x, n, ..., dst_nonexistent = NULL, dst_ambiguous = NULL)

# S3 method for Date
subtract_weeks(x, n, ...)

# S3 method for POSIXt
subtract_weeks(x, n, ..., dst_nonexistent = NULL, dst_ambiguous = NULL)

# S3 method for civil_zoned
subtract_weeks(x, n, ..., dst_nonexistent = NULL, dst_ambiguous = NULL)

# S3 method for Date
add_days(x, n, ...)

# S3 method for POSIXt
add_days(x, n, ..., dst_nonexistent = NULL, dst_ambiguous = NULL)

# S3 method for civil_zoned
add_days(x, n, ..., dst_nonexistent = NULL, dst_ambiguous = NULL)

# S3 method for Date
subtract_days(x, n, ...)

# S3 method for POSIXt
subtract_days(x, n, ..., dst_nonexistent = NULL, dst_ambiguous = NULL)

# S3 method for civil_zoned
subtract_days(x, n, ..., dst_nonexistent = NULL, dst_ambiguous = NULL)

# S3 method for Date
add_hours(x, n, ...)

# S3 method for POSIXt
add_hours(x, n, ...)

# S3 method for civil_zoned
add_hours(x, n, ...)

# S3 method for Date
subtract_hours(x, n, ...)

# S3 method for POSIXt
subtract_hours(x, n, ...)

# S3 method for civil_zoned
subtract_hours(x, n, ...)

# S3 method for Date
add_minutes(x, n, ...)

# S3 method for POSIXt
add_minutes(x, n, ...)

# S3 method for civil_zoned
add_minutes(x, n, ...)

# S3 method for Date
subtract_minutes(x, n, ...)

# S3 method for POSIXt
subtract_minutes(x, n, ...)

# S3 method for civil_zoned
subtract_minutes(x, n, ...)

# S3 method for Date
add_seconds(x, n, ...)

# S3 method for POSIXt
add_seconds(x, n, ...)

# S3 method for civil_zoned
add_seconds(x, n, ...)

# S3 method for Date
subtract_seconds(x, n, ...)

# S3 method for POSIXt
subtract_seconds(x, n, ...)

# S3 method for civil_zoned
subtract_seconds(x, n, ...)

# S3 method for Date
add_milliseconds(x, n, ...)

# S3 method for POSIXt
add_milliseconds(x, n, ...)

# S3 method for civil_zoned
add_milliseconds(x, n, ...)

# S3 method for Date
subtract_milliseconds(x, n, ...)

# S3 method for POSIXt
subtract_milliseconds(x, n, ...)

# S3 method for civil_zoned
subtract_milliseconds(x, n, ...)

# S3 method for Date
add_microseconds(x, n, ...)

# S3 method for POSIXt
add_microseconds(x, n, ...)

# S3 method for civil_zoned
add_microseconds(x, n, ...)

# S3 method for Date
subtract_microseconds(x, n, ...)

# S3 method for POSIXt
subtract_microseconds(x, n, ...)

# S3 method for civil_zoned
subtract_microseconds(x, n, ...)

# S3 method for Date
add_nanoseconds(x, n, ...)

# S3 method for POSIXt
add_nanoseconds(x, n, ...)

# S3 method for civil_zoned
add_nanoseconds(x, n, ...)

# S3 method for Date
subtract_nanoseconds(x, n, ...)

# S3 method for POSIXt
subtract_nanoseconds(x, n, ...)

# S3 method for civil_zoned
subtract_nanoseconds(x, n, ...)

Arguments

x
[Date / POSIXct / POSIXlt]

A zoned date-time vector.

n
[integer]

An integer vector representing the number of units to add to or subtract from x.

...
These dots are for future extensions and must be empty.

day_nonexistent
[character(1)]

Control the behavior when a nonexistent day is generated.

"last-time": Adjust to the last possible time of the current month.

"first-time": Adjust to the first possible time of the following month.

"last-day": Adjust to the last day of the current month. For date-times, the sub-daily components are kept.

"first-day": Adjust to the first day of the following month. For date-times, the sub-daily components are kept.

"NA": Replace the nonexistent date with NA.

"error": Error on nonexistent dates.

Warning: When used in arithmetic with date-times, "last-day" and "first-day" do not guarantee that the relative ordering of x is maintained.

dst_nonexistent
[NULL / character]

Control the behavior when a nonexistent time is generated due to a daylight savings gap.

NULL: If n is positive, choose "roll-forward". If n is negative, choose "roll-backward".

"roll-forward": Roll forward to the next valid moment in time.

"roll-backward": Roll backward to the previous valid moment in time.

"shift-forward": Shift the nonexistent time forward by the length of the daylight savings gap (which is usually 1 hour).

"shift-backward": Shift the nonexistent time backward by the length of the daylight savings gap (which is usually 1 hour).

"NA": Replace the nonexistent time with NA.

error: Error on nonexistent times.

Warning: When used in arithmetic, "shift-directional", "shift-forward", and "shift-backward" do not guarantee that the relative ordering of x is maintained.

dst_ambiguous
[NULL / character]

Control the behavior when an ambiguous time is generated due to a daylight savings fallback.

NULL: If n is positive, choose "earliest". If n is negative, choose "latest".

"earliest": Choose the earliest of the two possible ambiguous times.

"latest": Choose the latest of the two possible ambiguous times.

"NA": Replace the ambiguous time with NA.

"error": Error on ambiguous times.

Examples

x <- as.Date("2019-01-31")

add_days(x, 2)

#> [1] "2019-02-02"

# Adding one month theoretically lands us on
# "2019-02-31"
# but this doesn't exist, so `day_nonexistent`
# is consulted
add_months(x, 1)

#> [1] "2019-02-28"

add_months(x, 1, day_nonexistent = "first-time")

#> [1] "2019-03-01"

x <- as.POSIXct("1970-04-25 02:30:00", "America/New_York")
y <- as.POSIXct("1970-04-27 02:30:00", "America/New_York")

# Adding 1 day to `x` results in the nonexistent time of:
# "1970-04-26 02:30:00"
# at which point we have to make a decision about how to proceed using
# `dst_nonexistent`.
#
# Since we are adding a positive number of days, the default
# chooses `"roll-forward"` which rolls forward to the
# next valid moment in time.
add_days(x, 1)

#> [1] "1970-04-26 03:00:00 EDT"

# If we approach from the other side of the gap and subtract 1 day, then
# the default chooses `"roll-backward"`, which rolls backward
# to the previous valid moment in time.
subtract_days(y, 1)

#> [1] "1970-04-26 01:59:59 EST"

# If you want to force one of these options,
# you can set `dst_nonexistent` directly
add_days(x, 1, dst_nonexistent = "roll-backward")

#> [1] "1970-04-26 01:59:59 EST"

add_days(x, 1, dst_nonexistent = "NA")

#> [1] NA